@@ -55,49 +55,49 @@
     "@angular/platform-browser": ">5.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.6",
-    "@angular/animations": "^6.1.10",
-    "@angular/cli": "^6.2.6",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
-    "@types/jasmine": "^2.8.9",
-    "@types/jasminewd2": "^2.0.5",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/animations": "~8.0.1",
+    "@angular/cli": "~8.0.3",
+    "@angular/common": "~8.0.1",
+    "@angular/compiler": "~8.0.1",
+    "@angular/compiler-cli": "~8.0.1",
+    "@angular/core": "~8.0.1",
+    "@angular/forms": "~8.0.1",
+    "@angular/language-service": "~8.0.1",
+    "@angular/platform-browser": "~8.0.1",
+    "@angular/platform-browser-dynamic": "~8.0.1",
+    "@angular/router": "~8.0.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
     "@types/marked": "^0.3.0",
     "@types/node": "~8.9.4",
-    "@types/webpack": "^3.8.17",
+    "@types/webpack": "^4.4.34",
     "codecov": "^3.0.0",
-    "codelyzer": "~4.3.0",
+    "codelyzer": "^5.0.0",
     "core-js": "^2.5.4",
     "escape-string-regexp": "^1.0.5",
     "html-loader": "^0.5.5",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "lodash": "^4.17.11",
-    "markdown-loader": "^2.0.2",
+    "markdown-loader": "^5.0.0",
     "ng-packagr": "^4.4.0",
-    "ngx-bootstrap": "^3.1.1",
+    "ngx-bootstrap": "^5.0.0",
     "pre-commit": "^1.2.2",
     "protractor": "~5.4.0",
-    "rxjs": "^6.3.3",
+    "rxjs": "~6.4.0",
     "standard-version": "^4.3.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~2.9.2",
-    "zone.js": "~0.8.26"
+    "tslib": "^1.9.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3",
+    "zone.js": "~0.9.1"
   },
   "contributors": [
     {
